Deal with multi-byte ByteString to Char
from Philip Weaver:

In the instance Convertible SqlValue Char, in the case of a
SqlByteString, the code looks like this:

  ...

This will fail if the ByteString corresponds to a single multi-byte
unicode character.  B.length returns the number of *bytes* in the
ByteString, not the number of unicode characters.  I believe this case
should just skip the check, and simply do this:

...

After we've decoded the ByteString to a String, a single multibyte
character will correctly be a singleton String.